Move Styled components to Styled Folder.
