+++ src/components/Styled/StyledSearchBar.js
